1. UFUNCTION(BlueprintImplementableEvent) A mrco to create a event that can notified the Blue Print To call it use the event name and parameter to return;
2.
Using ensure marco to doing assertion. Supper useful for debugging
 if (!ensure(LeftTrack && RightTrack)) return;
ensure mean if will go inside the brecket, and check whether argument are true or not, if ture, ack like normal. If false provide lot informations

3. refactory Strucuture
Refactory Structure so that there is not spagatti code and strucutre.
Decoupling and cohesion is important